Make toolbarHelp as a component
