+++ src/components/map/toolbar/toolbar.scss
@@ -23,15 +23,4 @@
             }
         }
     }
-
-    .toolbarSubMenu {
-        display: flex;
-        background-color: $white;
-        border-radius: 3px;
-        box-shadow: 0 0 4px $grey;
-        min-height: 50px;
-        padding: 10px;
-        max-width: 254px;
-        font-size: 1rem;
-    }
 }
