+++ src/components/map/toolbar/toolbarHelp.scss
@@ -1,0 +1,12 @@
+@import '../../../common.scss';
+
+.toolbarHelp {
+  display: flex;
+  background-color: $white;
+  border-radius: 3px;
+  box-shadow: 0 0 4px $grey;
+  min-height: 50px;
+  padding: 10px;
+  max-width: 254px;
+  font-size: 1rem;
+}
